Remove H2 and Postgres storage backends
