--- src/packaging/resource/cassandra-reaper-postgres.yaml
@@ -1,134 +1,0 @@
-# Copyright 2015-2017 Spotify AB
-# Copyright 2016-2019 The Last Pickle Ltd
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Reaper for Apache Cassandra Configuration Example.
-# See a bit more complete example in:
-# src/server/src/test/resources/cassandra-reaper.yaml
-segmentCountPerNode: 64
-repairParallelism: DATACENTER_AWARE
-repairIntensity: 0.9
-scheduleDaysBetween: 7
-repairRunThreadCount: 15
-hangingRepairTimeoutMins: 30
-storageType: postgres
-enableCrossOrigin: true
-incrementalRepair: false
-blacklistTwcsTables: true
-enableDynamicSeedList: true
-repairManagerSchedulingIntervalSeconds: 10
-jmxConnectionTimeoutInSeconds: 5
-useAddressTranslator: false
-maxParallelRepairs: 2
-# purgeRecordsAfterInDays: 30
-# numberOfRunsToKeepPerUnit: 10
-
-# datacenterAvailability has three possible values: ALL | LOCAL | EACH | SIDECAR
-# the correct value to use depends on whether jmx ports to C* nodes in remote datacenters are accessible
-# If the reaper has access to all node jmx ports, across all datacenters, then configure to ALL.
-# If jmx access is only available to nodes in the same datacenter as reaper in running in, then configure to LOCAL.
-# If there's a reaper instance running in every datacenter, and it's important that nodes under duress are not involved in repairs,
-#    then configure to EACH.
-# If jmx access is restricted to localhost, then configure to SIDECAR.
-#
-# The default is ALL
-datacenterAvailability: ALL
-
-
-# any cassandra nodes on 127.0.0.x addresses we presume are ccm nodes
-jmxPorts:
-  127.0.0.1: 7100
-  127.0.0.2: 7200
-  127.0.0.3: 7300
-  127.0.0.4: 7400
-  127.0.0.5: 7500
-  127.0.0.6: 7600
-  127.0.0.7: 7700
-  127.0.0.8: 7800
-
-#jmxAuth:
-#  username: myUsername
-#  password: myPassword
-
-logging:
-  level: INFO
-  loggers:
-    io.dropwizard: WARN
-    org.eclipse.jetty: WARN
-  appenders:
-    - type: console
-      logFormat: "%-6level [%d] [%t] %logger{5} - %msg %n"
-      threshold: WARN
-    - type: file
-      logFormat: "%-6level [%d] [%t] %logger{5} - %msg %n"
-      currentLogFilename: /var/log/cassandra-reaper/reaper.log
-      archivedLogFilenamePattern: /var/log/cassandra-reaper/reaper-%d.log.gz
-      archivedFileCount: 99
-
-server:
-  type: default
-  applicationConnectors:
-    - type: http
-      port: 8080
-      bindHost: 0.0.0.0
-  adminConnectors:
-    - type: http
-      port: 8081
-      bindHost: 0.0.0.0
-  requestLog:
-    appenders: []
-
-postgres:
-  driverClass: org.postgresql.Driver
-  user: postgres
-  password: 
-  url: jdbc:postgresql://127.0.0.1/reaper
-
-autoScheduling:
-  enabled: false
-  initialDelayPeriod: PT15S
-  periodBetweenPolls: PT10M
-  timeBeforeFirstSchedule: PT5M
-  scheduleSpreadPeriod: PT6H
-  excludedKeyspaces:
-    - keyspace1
-    - keyspace2
-  excludedClusters:
-#   - cluster1
-
-
-# Uncomment the following to enable dropwizard metrics
-#  Configure to the reporter of your choice
-#  Reaper also provides prometheus metrics on the admin port at /prometheusMetrics
-
-#metrics:
-#  frequency: 1 minute
-#  reporters:
-#    - type: log
-#      logger: metrics
-
-# Uncomment the following to enable encryption and the ability to store jmx credentials in
-#  the backend data source where SOME_SYSTEM_PROPERTY_KEY is a system property that can be
-#  used to lookup the shared symmetric key.
-#  ie: export SOME_SYSTEM_PROPERTY=YOUR_SHARED_SYMMETRIC_KEY
-
-#cryptograph:
-#  type: symmetric
-#  systemPropertySecret: SOME_SYSTEM_PROPERTY
-
-# Authentication is enabled by default
-accessControl:
-  sessionTimeout: PT10M
-  shiro:
-    iniConfigs: ["classpath:shiro.ini"]
