--- src/server/src/main/java/io/cassandrareaper/storage/postgresql/InstantArgumentFactory.java
@@ -1,43 +1,0 @@
-/*
- *
- * Copyright 2019-2019 The Last Pickle Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.cassandrareaper.storage.postgresql;
-
-import java.sql.Timestamp;
-import java.time.Instant;
-
-import org.skife.jdbi.v2.StatementContext;
-import org.skife.jdbi.v2.tweak.Argument;
-import org.skife.jdbi.v2.tweak.ArgumentFactory;
-
-public class InstantArgumentFactory implements ArgumentFactory<Instant> {
-
-  @Override
-  public boolean accepts(Class<?> expectedType, Object value, StatementContext ctx) {
-    return value instanceof Instant;
-  }
-
-  @Override
-  public Argument build(Class<?> expectedType, final Instant value, StatementContext ctx) {
-    return (pos, stmt, sc) -> stmt.setTimestamp(pos, toSqlTimestamp(value));
-  }
-
-  private static Timestamp toSqlTimestamp(Instant instant) {
-    return Timestamp.from(instant);
-  }
-
-}
