+++ src/server/src/test/java/io/cassandrareaper/storage/DiagEventSubscriptionMapper.java
@@ -1,6 +1,7 @@
 /*
  * Copyright 2018-2018 Stefan Podkowinski
  * Copyright 2019-2019 The Last Pickle Ltd
+ * Copyright 2021-2021 DataStax, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.cassandrareaper.storage.postgresql;
+package io.cassandrareaper.storage;
 
 import io.cassandrareaper.core.DiagEventSubscription;
