+++ src/server/src/test/resources/reaper-cassandra-each1-at.yaml
@@ -59,13 +59,6 @@ jmxCredentials:
     username: cassandra
     password: cassandrapassword
 
-# database section will be ignored if storageType is set to "cassandra"
-h2:
-  driverClass: org.h2.Driver
-  url: jdbc:h2:mem:reaper-db;MODE=PostgreSQL
-  user:
-  password:
-  
 cassandra:
   clusterName: "test"
   contactPoints: ["127.0.0.1"]
