--- src/server/src/test/resources/reaper-postgres-each2-at.yaml
@@ -1,75 +1,0 @@
-# Copyright 2017-2017 Spotify AB
-# Copyright 2017-2019 The Last Pickle Ltd
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Reaper for Apache Cassandra Configuration for Acceptance Tests.
-#
-segmentCountPerNode: 16
-repairParallelism: SEQUENTIAL
-repairIntensity: 0.95
-scheduleDaysBetween: 7
-repairRunThreadCount: 15
-hangingRepairTimeoutMins: 1
-storageType: postgres
-incrementalRepair: false
-blacklistTwcsTables: true
-jmxConnectionTimeoutInSeconds: 10
-activateQueryLogger: true
-datacenterAvailability: EACH
-enableConcurrentMigrations: false
-percentRepairedCheckIntervalMinutes: 1
-
-logging:
-  level: WARN
-  appenders:
-    - type: console
-
-server:
-  type: default
-  applicationConnectors:
-    - type: http
-      port: 8083
-      bindHost: 127.0.0.1
-  adminConnectors:
-    - type: http
-      port: 8084
-      bindHost: 127.0.0.1
-
-jmxPorts:
-  127.0.0.1: 7999
-  127.0.0.2: 7999
-  127.0.0.3: 7300
-  127.0.0.4: 7400
-
-jmxCredentials:
-  test:
-    username: cassandra
-    password: cassandrapassword
-
-# database section will be ignored if storageType is set to "memory"
-postgres:
-  driverClass: org.postgresql.Driver
-  user: postgres
-  password: postgres
-  url: jdbc:postgresql://127.0.0.1/reaper
-
-metrics:
-  frequency: 1 second
-  reporters:
-    - type: csv
-      file: target/dropwizard-metrics
-
-cryptograph:
-  type: symmetric
-  systemPropertySecret: REAPER_ENCRYPTION_KEY
